+++ pkg/collector/corechecks/oracle/cache_wrapper.go
