+++ pkg/collector/corechecks/oracle/common/utils.go
