+++ pkg/collector/corechecks/oracle/config/metric.go
