+++ pkg/collector/corechecks/oracle/plan_queries.go
