+++ pkg/collector/corechecks/oracle/statement_metrics_queries.go
