+++ pkg/collector/corechecks/oracle/testutil.go
@@ -38,7 +38,7 @@ var USER = "c##datadog"
 var PASSWORD = "datadog"
 var SERVICE_NAME = "XE"
 var TNS_ALIAS = "XE"
-var TNS_ADMIN = "/Users/nenad.noveljic/go/src/github.com/DataDog/datadog-agent/pkg/collector/corechecks/oracle-dbm/testutil/etc/netadmin"
+var TNS_ADMIN = "/Users/nenad.noveljic/go/src/github.com/DataDog/datadog-agent/pkg/collector/corechecks/oracle/testutil/etc/netadmin"
 
 var dbmsTag = "dbms:oracle"
