+++ tasks/agent.py
@@ -67,7 +67,7 @@ AGENT_CORECHECKS = [
     "memory",
     "ntp",
     "oom_kill",
-    "oracle-dbm",
+    "oracle",
     "sbom",
     "systemd",
     "tcp_queue_length",
